refactor(kernel): relocate non-foundational stragglers before packaging
W-pre of #1490: kernel's file set becomes its public API before it
moves to packages/. Three files drifted into the sink because R1
pushes shared code down, not because they are foundational:

- audio-probe-support -> src/contracts/ (capability policy consumed by
  platform plugins and interactor registration)
- input-validation -> src/contracts/ (lowest-rank consumer is
  contracts/gesture-input)
- observability-redaction -> src/commands/observability/runtime/
  (colocated with its only consumer, diagnostics-format)

Pure moves; no behavior change. Kernel retains errors, device,
snapshot, contracts, collections, rect, redaction, daemon-error.
