+++ src/contracts/input-validation.ts
@@ -1,4 +1,4 @@
-import { AppError } from './errors.ts';
+import { AppError } from '../kernel/errors.ts';
 
 export function readOptionalInteger(
   record: Record<string, unknown>,
